Create LexicalAnalyser.c
Tokens are the smallest elements of a program, which are meaningful to the compiler. 
The following are the types of tokens: Keywords, Identifiers, Constants, Strings, Operators, etc. This is called the lexical analysis phase of the compiler. 
The lexical analyzer is the part of the compiler that detects the token of the program and sends it to the syntax analyzer.